Clearly state that GeoJSON is a RFC6906 profile of JSON.
RFC4267 doesn't specify any media type parameters, but a profile
parameter is a big help to GeoJSON. Without it, we'd probably need
a custom media type. GeoJSON's use of a profile parameter does not
alter the (JSON) processing model, it only indicates that the root
object contains geographic features to be interpreted according to
the GeoJSON spec.